+++ Nos/Assets/Localization/Localized.swift
@@ -55,10 +55,7 @@ enum Localized: String, Localizable, CaseIterable {
     case error = "Error"
     case noEvents = "No notes to show! Add a relay to get started."
     case addItem = "Add Item"
-    case homeFeed = "Home Feed"
-    case homeFeedLinkTitle = "🏠 Home Feed"
-    case relaysLinkTitle = "📡 Relays"
-    case settingsLinkTitle = "⚙️ Settings"
+    case homeFeed = "Feed"
     
     case keys = "Keys"
     case keyEncryptionWarning = "Warning: Never share your private key with anyone."
@@ -79,6 +76,7 @@ enum Localized: String, Localizable, CaseIterable {
     
     case profile = "Profile"
     case follow = "Follow"
+    case follows = "Follows"
     case following = "Following"
     case unfollow = "Unfollow"
     case readMore = "Read more"
@@ -95,8 +93,11 @@ enum Localized: String, Localizable, CaseIterable {
     case noNotifications = "No notifications (yet)!"
     case copyNoteIdentifier = "Copy Note Identifier"
     case copyUserIdentifier = "Copy User ID"
+    case muteUser = "Mute"
+    case mutedUser = "Muted"
     case share = "Share"
     case reportPost = "Report this post"
+    case unmuteUser = "Un-Mute"
 }
 
 // MARK: - Onboarding