+++ Nos/Views/ProfileHeader.swift
@@ -66,12 +66,19 @@ struct ProfileHeader: View {
                                 .font(.title3.weight(.semibold))
                                 .foregroundColor(Color.primaryTxt)
                             Spacer()
-                            FollowButton(currentUserAuthor: CurrentUser.author, author: author)
+                            if let currentUser = CurrentUser.author {
+                                FollowButton(currentUserAuthor: currentUser, author: author)
+                                if author.muted {
+                                    Text(Localized.mutedUser.string)
+                                        .font(.subheadline)
+                                        .foregroundColor(Color.secondaryTxt)
+                                }
+                            }
                         }
                         Spacer()
 
                         Button {
-                            router.path.append(follows)
+                            router.currentPath.wrappedValue.append(follows)
                         } label: {
                             Text("\(Localized.following.string): \(author.follows?.count ?? 0)")
                         }