+++ Nos/Views/ProfileView.swift
@@ -11,7 +11,7 @@ import CoreData
 struct ProfileView: View {
     @EnvironmentObject var router: Router
     
-    var author: Author
+    @ObservedObject var author: Author
     
     @Environment(\.managedObjectContext) private var viewContext
     
@@ -34,7 +34,7 @@ struct ProfileView: View {
             subscriptionIds.removeAll()
         }
 
-        if let currentUserPublicKey = CurrentUser.author.hexadecimalPublicKey {
+        if let currentUserPublicKey = CurrentUser.author?.hexadecimalPublicKey {
             let authors = [currentUserPublicKey]
             let textFilter = Filter(authorKeys: authors, kinds: [.text], limit: 100)
             let textSub = relayService.requestEventsFromAll(filter: textFilter)
@@ -58,7 +58,7 @@ struct ProfileView: View {
                     .shadow(color: .profileShadow, radius: 10, x: 0, y: 4)
                 
                 LazyVStack {
-                    ForEach(events) { event in
+                    ForEach(events.unmuted) { event in
                         VStack {
                             NoteButton(note: event)
                                 .padding(.horizontal)
@@ -69,16 +69,15 @@ struct ProfileView: View {
             .padding(.top, 1)
             .background(Color.appBg)
             .overlay(Group {
-                if events.isEmpty {
+                if !events.contains(where: { !$0.author!.muted }) {
                     Localized.noEventsOnProfile.view
                         .padding()
                 }
             })
         }
-        .navigationBarBackButtonHidden(true)
+        .navigationBarTitle(Localized.profile.rawValue, displayMode: .inline)
         .onAppear {
-            router.navigationTitle = Localized.profile.rawValue
-            router.userNpubPublicKey = author.publicKey?.npub ?? ""
+            router.viewedAuthor = author
         }
         .task {
             refreshProfileFeed()