+++ Nos/Views/RelayView.swift
@@ -21,7 +21,7 @@ struct RelayView: View {
     @Dependency(\.analytics) private var analytics
     
     var body: some View {
-        NavigationStack(path: $router.path) {
+        NavigationStack(path: $router.relayPath) {
             List {
                 if let relays = author.relays?.allObjects as? [Relay] {
                     Section(Localized.relays.string) {
@@ -48,6 +48,8 @@ struct RelayView: View {
                             }
 
                             try! viewContext.save()
+                            
+                            publishChanges()
                         }
 
                         if author.relays?.count == 0 {
@@ -65,6 +67,7 @@ struct RelayView: View {
                     Button(Localized.save.string) {
                         addRelay()
                         CurrentUser.subscribe()
+                        publishChanges()
                     }
                 }
                 if author.relays?.count == 0 {
@@ -82,13 +85,13 @@ struct RelayView: View {
             .navigationDestination(for: Author.self) { author in
                 ProfileView(author: author)
             }
-            .navigationDestination(for: AppView.Destination.self) { destination in
-                if destination == AppView.Destination.settings {
-                    SettingsView()
-                }
-            }
         }
         .navigationTitle(Localized.relays.string)
+    }
+    
+    func publishChanges() {
+        let followKeys = CurrentUser.follows?.keys ?? []
+        CurrentUser.publishContactList(tags: followKeys.tags, context: viewContext)
     }
     
     private func addRelay() {
@@ -100,7 +103,7 @@ struct RelayView: View {
             relay.createdAt = Date.now
             newRelayAddress = ""
 
-            CurrentUser.author.add(relay: relay)
+            CurrentUser.author?.add(relay: relay)
             
             do {
                 try viewContext.save()
@@ -121,14 +124,20 @@ struct RelayView_Previews: PreviewProvider {
     static var previewContext = PersistenceController.preview.container.viewContext
     
     static var emptyContext = PersistenceController.empty.container.viewContext
+
+    static var user: Author {
+        let author = Author(context: previewContext)
+        author.hexadecimalPublicKey = "d0a1ffb8761b974cec4a3be8cbcb2e96a7090dcf465ffeac839aa4ca20c9a59e"
+        return author
+    }
     
     static var previews: some View {
         NavigationStack {
-            RelayView(author: CurrentUser.author)
+            RelayView(author: user)
         }.environment(\.managedObjectContext, previewContext)
         
         NavigationStack {
-            RelayView(author: CurrentUser.author)
+            RelayView(author: user)
         }.environment(\.managedObjectContext, emptyContext)
     }
 }