+++ Nos/Views/ThreadView.swift
@@ -34,7 +34,8 @@ struct ThreadView: View {
     
     var body: some View {
         LazyVStack {
-            NoteButton(note: root)
+            NoteButton(note: root, showFullMessage: true)
+                .disabled(true)
                 .padding(.horizontal)
             ForEach(thread) { event in
                 VStack {